# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentCreateRequest": {
        "additionalProperties": false,
        "description": "由模板新增一名本场角色。",
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 30,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "template_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Template Id",
            "type": "string"
          }
        },
        "required": [
          "template_id"
        ],
        "title": "AgentCreateRequest",
        "type": "object"
      },
...
      "AgentRenameRequest": {
        "additionalProperties": false,
        "description": "重命名本场角色（显示名称）。",
        "properties": {
          "name": {
            "maxLength": 30,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "AgentRenameRequest",
        "type": "object"
      },
...
      "AgentSpecRequest": {
        "additionalProperties": false,
        "description": "创建场景时对一名本场角色的要求。",
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 30,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "template_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Template Id",
            "type": "string"
          }
        },
        "required": [
          "template_id"
        ],
        "title": "AgentSpecRequest",
        "type": "object"
      },
...
      "AgentTemplate": {
        "additionalProperties": false,
        "description": "角色模板；支持创建、编辑、复制、列表查看与移除（PRD 3.2）。",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "initial_goal": {
            "title": "Initial Goal",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "persona": {
            "title": "Persona",
            "type": "string"
          },
          "private_background": {
            "title": "Private Background",
            "type": "string"
          },
          "speech_style": {
            "title": "Speech Style",
            "type": "string"
          },
          "template_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Template Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "persona",
          "speech_style",
          "initial_goal",
          "private_background",
          "template_id",
          "created_at",
          "updated_at"
        ],
        "title": "AgentTemplate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PauseReason": {
        "description": "暂停原因必须区分（PRD 5.2）。",
        "enum": [
          "NO_NEW_INFORMATION",
          "MANUAL",
          "PROVIDER_ERROR",
          "CONTEXT_LIMIT",
          "PROCESS_INTERRUPT"
        ],
        "title": "PauseReason",
        "type": "string"
      },
...
      "PresetSceneCreateRequest": {
        "additionalProperties": false,
        "description": "用预置场景创建会话。",
        "properties": {
          "preset_key": {
            "default": "roommates",
            "maxLength": 64,
            "minLength": 1,
            "title": "Preset Key",
            "type": "string"
          }
        },
        "title": "PresetSceneCreateRequest",
        "type": "object"
      },
...
      "RunState": {
        "description": "会话运行状态（PRD 5.2）。没有其他状态。",
        "enum": [
          "READY",
          "RUNNING",
          "PAUSING",
          "STOPPING",
          "PAUSED",
          "ENDED"
        ],
        "title": "RunState",
        "type": "string"
      },
...
      "Scene": {
        "additionalProperties": false,
        "description": "场景／会话（PRD 3.1、5.2）。一个活动场景，多份历史记录。",
        "properties": {
          "background": {
            "title": "Background",
            "type": "string"
          },
          "budget": {
            "$ref": "#/components/schemas/Budget"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "pause_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PauseReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "scene_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scene Id",
            "type": "string"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunState",
            "default": "READY"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "scene_id",
          "title",
          "background",
          "budget",
          "created_at"
        ],
        "title": "Scene",
        "type": "object"
      },
      "SceneAgent": {
        "additionalProperties": false,
        "description": "本场角色实例：``agent_id`` 是关联主键，``name`` 是显示字段。",
        "properties": {
          "agent_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "order_index": {
            "minimum": 0.0,
            "title": "Order Index",
            "type": "integer"
          },
          "scene_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scene Id",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/AgentSnapshot"
          }
        },
        "required": [
          "agent_id",
          "scene_id",
          "name",
          "order_index",
          "snapshot",
          "created_at"
        ],
        "title": "SceneAgent",
        "type": "object"
      },
      "SceneCreateRequest": {
        "additionalProperties": false,
        "description": "创建场景（会话）。\n\n``agents`` 的 2～8 人约束在契约层即校验，因此越界请求得到 422 而不是\n进入业务层（PRD 1.2、9）。",
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentSpecRequest"
            },
            "maxItems": 8,
            "minItems": 2,
            "title": "Agents",
            "type": "array"
          },
          "background": {
            "default": "",
            "maxLength": 2000,
            "title": "Background",
            "type": "string"
          },
          "max_analysis_requests": {
            "anyOf": [
              {
                "maximum": 200.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Analysis Requests"
          },
          "max_role_requests": {
            "anyOf": [
              {
                "maximum": 200.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Role Requests"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "agents"
        ],
        "title": "SceneCreateRequest",
        "type": "object"
      },
      "SceneDetailView": {
        "additionalProperties": false,
        "description": "场景详情：场景 + 本场角色集合（含模板快照）。",
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/SceneAgent"
            },
            "title": "Agents",
            "type": "array"
          },
          "locked": {
            "title": "Locked",
            "type": "boolean"
          },
          "scene": {
            "$ref": "#/components/schemas/Scene"
          }
        },
        "required": [
          "scene",
          "agents",
          "locked"
        ],
        "title": "SceneDetailView",
        "type": "object"
      },
      "SceneListView": {
        "additionalProperties": false,
        "properties": {
          "scenes": {
            "items": {
              "$ref": "#/components/schemas/SceneSummaryView"
            },
            "title": "Scenes",
            "type": "array"
          }
        },
        "required": [
          "scenes"
        ],
        "title": "SceneListView",
        "type": "object"
      },
      "SceneSummaryView": {
        "additionalProperties": false,
        "description": "场景列表项。",
        "properties": {
          "agent_count": {
            "minimum": 0.0,
            "title": "Agent Count",
            "type": "integer"
          },
          "budget_locked": {
            "title": "Budget Locked",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "scene_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scene Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "scene_id",
          "title",
          "status",
          "agent_count",
          "budget_locked",
          "created_at"
        ],
        "title": "SceneSummaryView",
        "type": "object"
      },
...
      "TemplateCopyRequest": {
        "additionalProperties": false,
        "description": "复制角色模板；省略 ``name`` 时自动取第一个可用的不重复名称。",
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 30,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "TemplateCopyRequest",
        "type": "object"
      },
      "TemplateCreateRequest": {
        "additionalProperties": false,
        "description": "创建角色模板。",
        "properties": {
          "initial_goal": {
            "title": "Initial Goal",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "persona": {
            "title": "Persona",
            "type": "string"
          },
          "private_background": {
            "title": "Private Background",
            "type": "string"
          },
          "speech_style": {
            "title": "Speech Style",
            "type": "string"
          }
        },
        "required": [
          "name",
          "persona",
          "speech_style",
          "initial_goal",
          "private_background"
        ],
        "title": "TemplateCreateRequest",
        "type": "object"
      },
      "TemplateListView": {
        "additionalProperties": false,
        "description": "模板列表。",
        "properties": {
          "templates": {
            "items": {
              "$ref": "#/components/schemas/AgentTemplate"
            },
            "title": "Templates",
            "type": "array"
          }
        },
        "required": [
          "templates"
        ],
        "title": "TemplateListView",
        "type": "object"
      },
      "TemplateUpdateRequest": {
        "additionalProperties": false,
        "description": "编辑角色模板（部分字段）。``None`` 表示不修改。",
        "properties": {
          "initial_goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Goal"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 30,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "private_background": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Private Background"
          },
          "speech_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speech Style"
          }
        },
        "title": "TemplateUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SceneWeave（角色互动沙盒）后端。M00 提供工程基础与公共契约；M01 提供角色模板与场景配置；M02/M03 提供可见性、调度与模型适配；M04 提供运行控制、事件、时间线、角色视角与 SSE；M06 提供只读行为分析。",
    "title": "SceneWeave",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scenes": {
      "get": {
        "operationId": "list_scenes_api_scenes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneListView"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Scenes",
        "tags": [
          "m01-scenes"
        ]
      },
      "post": {
        "operationId": "create_scene_api_scenes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scene",
        "tags": [
          "m01-scenes"
        ]
      }
    },
    "/api/scenes/preset": {
      "post": {
        "description": "用预置“三个室友的客厅”创建会话（PRD 3.1）。",
        "operationId": "create_preset_scene_api_scenes_preset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetSceneCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Preset Scene",
        "tags": [
          "m01-scenes"
        ]
      }
    },
...
    "/api/scenes/{scene_id}": {
      "get": {
        "operationId": "get_scene_api_scenes__scene_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scene",
        "tags": [
          "m01-scenes"
        ]
      }
    },
    "/api/scenes/{scene_id}/agents": {
      "post": {
        "operationId": "add_agent_api_scenes__scene_id__agents_post",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneAgent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Agent",
        "tags": [
          "m01-scenes"
        ]
      }
    },
...
    "/api/scenes/{scene_id}/agents/{agent_id}": {
      "delete": {
        "operationId": "remove_agent_api_scenes__scene_id__agents__agent_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Agent",
        "tags": [
          "m01-scenes"
        ]
      },
      "patch": {
        "operationId": "rename_agent_api_scenes__scene_id__agents__agent_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneAgent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Agent",
        "tags": [
          "m01-scenes"
        ]
      }
    },
...
    "/api/templates": {
      "get": {
        "operationId": "list_templates_api_templates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateListView"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Templates",
        "tags": [
          "m01-templates"
        ]
      },
      "post": {
        "operationId": "create_template_api_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTemplate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Template",
        "tags": [
          "m01-templates"
        ]
      }
    },
    "/api/templates/{template_id}": {
      "delete": {
        "operationId": "delete_template_api_templates__template_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Template",
        "tags": [
          "m01-templates"
        ]
      },
      "get": {
        "operationId": "get_template_api_templates__template_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTemplate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Template",
        "tags": [
          "m01-templates"
        ]
      },
      "patch": {
        "operationId": "update_template_api_templates__template_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTemplate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Template",
        "tags": [
          "m01-templates"
        ]
      }
    },
    "/api/templates/{template_id}/copy": {
      "post": {
        "operationId": "copy_template_api_templates__template_id__copy_post",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateCopyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTemplate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copy Template",
        "tags": [
          "m01-templates"
        ]
      }
    }
  }
}